+++ src/services/database.js
@@ -2,10 +2,9 @@ import { printPost } from '../components/feed.js'
 
 export const loadPosts = () => {
   return firebase.firestore().collection('posts')
-  // .orderBy('time', 'desc')
+  .orderBy('text', 'desc')
   .get()
   .then(snap => {
-    const posts = [];
     snap.forEach(post => {
       printPost(post);
     });
@@ -16,8 +15,12 @@ export const loadPosts = () => {
 }
 
 export const addPosts = (post) => {
-  firebase.firestore().collection('posts').add(post)
+  return firebase.firestore().collection('posts').add(post)
   .then(() => loadPosts());
+}
+
+export const updatePost = (valueInput) => {
+ return firebase.firestore().collection('posts').doc(post.id).update({ text: valueInput })
 }
 
 export const deletePost = (postId) => {
